Movido os testes de templatetags para o mudulo test_templatetags.py
